+++ apps/chaingraph-frontend/src/components/flow/nodes/ChaingraphNode/ports/ArrayPort/AddElementPopover.tsx
@@ -88,6 +88,7 @@ const typeConfigMap: Record<PortType, IPortConfig> = {
     },
     defaultValue: [],
     isMutable: true,
+    isSchemaMutable: true,
     ui: {
       hideEditor: false,
     },