+++ apps/chaingraph-frontend/src/components/flow/nodes/ChaingraphNode/ports/ArrayPort/EditItemConfigPopover.tsx
@@ -104,6 +104,7 @@ export function EditItemConfigPopover({
             title: 'Item',
           },
           isMutable: true,
+          isSchemaMutable: true,
         }
         break